Move loopback handling into its own module.
